Create common components for send view in sudo and console
