+++ frontend/src/lib/sends/MxHost.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { Tooltip } from '@hyvor/design/components';
-	import type { SmtpConversation } from '../../../types';
+	import type { SmtpConversation } from './types';
 
 	interface Props {
 		host: string;
