+++ frontend/src/lib/sends/RetryModal.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { Modal, Checkbox, TextInput, Radio } from '@hyvor/design/components';
-	import type { SendRecipient } from '../../../types';
+	import type { SendRecipient } from './types';
 
 	interface Props {
 		show: boolean;
