+++ frontend/src/routes/console/[id]/sends/[uuid]/+page.svelte
@@ -1,152 +1,15 @@
 <script lang="ts">
-	import { onMount } from 'svelte';
-	import {
-		Loader,
-		TabNav,
-		TabNavItem,
-		toast,
-		IconMessage,
-		Button,
-		CodeBlock
-	} from '@hyvor/design/components';
-	import { getEmailByUuid } from '../../../lib/actions/emailActions';
-	import type { Send } from '../../../types';
-	import SingleBox from '../../../@components/content/SingleBox.svelte';
-	import IconCaretLeft from '@hyvor/icons/IconCaretLeft';
-	import { consoleUrlProject } from '../../../lib/consoleUrl';
 	import { page } from '$app/state';
-	import Overview from './Overview.svelte';
-	import Preview from './Preview.svelte';
-
-	let send: Send | null = $state(null);
-	let loading = $state(true);
-	let error: string | null = $state(null);
-	let activeTab: 'overview' | 'preview' | 'raw' = $state('overview');
-
-	function fetchSend() {
-		const emailUuid = page.params.uuid ?? '';
-		return getEmailByUuid(emailUuid);
-	}
-
-	onMount(() => {
-		fetchSend()
-			.then((result) => {
-				send = result;
-			})
-			.catch((err: any) => {
-				error = err.message || 'Failed to load email';
-				toast.error(error);
-			})
-			.finally(() => {
-				loading = false;
-			});
-	});
+	import SingleBox from '../../../@components/content/SingleBox.svelte';
+	import { consoleUrlProject } from '../../../lib/consoleUrl';
+	import { getEmailByUuid, retrySend } from '../../../lib/actions/emailActions';
+	import SendDetail from '$lib/sends/SendDetail.svelte';
 </script>
 
 <SingleBox>
-	{#if loading}
-		<Loader full />
-	{:else if error}
-		<IconMessage error message={error} />
-	{:else if send}
-		<div class="email-detail">
-			<div class="header">
-				<div class="header-left">
-					<Button size="small" color="input" as="a" href={consoleUrlProject(`sends`)}>
-						{#snippet start()}
-							<IconCaretLeft size={12} />
-						{/snippet}
-						All Sends
-					</Button>
-				</div>
-			</div>
-
-			<div class="content">
-				<div class="tabs">
-					<TabNav>
-						<TabNavItem name="overview" active={activeTab === 'overview'} onclick={() => activeTab = 'overview'}>Overview</TabNavItem>
-						<TabNavItem name="preview" active={activeTab === 'preview'} onclick={() => activeTab = 'preview'}>Preview</TabNavItem>
-						<TabNavItem name="raw" active={activeTab === 'raw'} onclick={() => activeTab = 'raw'}>Raw</TabNavItem>
-					</TabNav>
-				</div>
-
-				{#if activeTab === 'overview'}
-					<Overview {send} onSendUpdate={(updated) => send = updated} />
-				{/if}
-
-				{#if activeTab === 'preview'}
-					<Preview {send} />
-				{/if}
-
-				{#if activeTab === 'raw'}
-					<div class="raw-content">
-						<div class="raw-content-note">
-							This is the raw email content, including headers and body.
-						</div>
-						<CodeBlock code={send.raw} language={null} />
-					</div>
-				{/if}
-			</div>
-		</div>
-	{/if}
+	<SendDetail
+		fetchSend={() => getEmailByUuid(page.params.uuid ?? '')}
+		backHref={consoleUrlProject('sends')}
+		onRetry={retrySend}
+	/>
 </SingleBox>
-
-<style>
-	.email-detail {
-		height: 100%;
-		display: flex;
-		flex-direction: column;
-	}
-
-	.header {
-		display: flex;
-		align-items: center;
-		justify-content: space-between;
-		padding: 15px 25px;
-		border-bottom: 1px solid var(--border);
-	}
-
-	.header-left {
-		display: flex;
-		align-items: center;
-		gap: 15px;
-	}
-
-	.content {
-		flex: 1;
-		overflow: auto;
-	}
-
-	.tabs {
-		padding: 20px 25px;
-	}
-
-	.raw-content {
-		flex: 1;
-		overflow: auto;
-		padding: 0px 25px;
-	}
-	.raw-content-note {
-		margin-bottom: 10px;
-		font-size: 14px;
-		color: var(--text-light);
-	}
-
-	@media (max-width: 768px) {
-		.email-detail {
-			padding: 15px;
-		}
-
-		.header {
-			flex-direction: column;
-			align-items: flex-start;
-			gap: 15px;
-		}
-
-		.header-left {
-			flex-direction: column;
-			align-items: flex-start;
-			gap: 10px;
-		}
-	}
-</style>
